refactor!: Move `waitForRemotePeer` to own file
