+++ src/index.ts
@@ -13,6 +13,8 @@ export * as enr from "./lib/enr";
 
 export * as utils from "./lib/utils";
 
+export { waitForRemotePeer } from "./lib/wait_for_remote_peer";
+
 export * as waku from "./lib/waku";
 export { Waku, Protocols } from "./lib/waku";
