+++ src/lib/waku_light_push/index.node.spec.ts
@@ -3,6 +3,7 @@ import debug from "debug";
 
 import { makeLogFileName, NOISE_KEY_1, Nwaku } from "../../test_utils";
 import { delay } from "../../test_utils/delay";
+import { waitForRemotePeer } from "../wait_for_remote_peer";
 import { createWaku, Protocols, Waku } from "../waku";
 import { WakuMessage } from "../waku_message";
 
@@ -30,7 +31,7 @@ describe("Waku Light Push [node only]", () => {
     });
     await waku.start();
     await waku.dial(await nwaku.getMultiaddrWithId());
-    await waku.waitForRemotePeer([Protocols.LightPush]);
+    await waitForRemotePeer(waku, [Protocols.LightPush]);
 
     const messageText = "Light Push works!";
     const message = await WakuMessage.fromUtf8String(
@@ -67,7 +68,7 @@ describe("Waku Light Push [node only]", () => {
     });
     await waku.start();
     await waku.dial(await nwaku.getMultiaddrWithId());
-    await waku.waitForRemotePeer([Protocols.LightPush]);
+    await waitForRemotePeer(waku, [Protocols.LightPush]);
 
     const nimPeerId = await nwaku.getPeerId();
